# web/static/swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sentinel Incident Viewer API",
        "description": "Interactive API guide for Sentinel. Public endpoints are open; internal endpoints require an authenticated session (developer/admin).\\n\\n### Webhook HMAC\\nFor webhook endpoints (generic/Slack/Discord/bot), include `X-Sentinel-Signature` computed as hex(HMAC-SHA256(body, secret)). Example in bash:\\n```bash\\nbody='{\"title\":\"Hello\",\"message\":\"World\"}'\\nsecret='YOUR_SECRET'\\nsig=$(printf \"%s\" \"$body\" | openssl dgst -sha256 -hmac \"$secret\" -binary | xxd -p -c 256)\\ncurl -X POST https://your-host/webhook \\\\\\n  -H \"X-Sentinel-Signature: $sig\" \\\\\\n  -H \"Content-Type: application/json\" \\\\\\n  -d \"$body\"\\n```\\nYou can also include a timestamp/nonce if you enforce it (e.g., add headers `X-Sentinel-Timestamp` and `X-Sentinel-Nonce` and sign `timestamp + '.' + nonce + '.' + body`).",
        "version": "1.0.0"
    },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie set after login"
      }
    },
    "schemas": {
      "Alert": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "message": { "type": "string" },
          "level": { "type": "string", "enum": ["info", "warning", "error", "critical", "success"] },
          "source": { "type": "string" }
        },
        "required": ["title", "message"]
      },
...
    }
  },
  "paths": {
...
    "/bot/{token}": {
      "post": {
        "tags": ["Public"],
        "summary": "Bot webhook (token keyed)",
        "parameters": [{ "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Alert" } } }
        },
        "responses": { "200": { "description": "Alert created" }, "404": { "description": "Invalid token" } }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "List users"
      },
      "post": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["admin", "developer", "user"] }
                },
                "required": ["username", "password", "role"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/users/{id}": {
      "put": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Update user",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "username": { "type": "string" }, "role": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Delete user",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
    "/api/admin/bots": {
      "get": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "List bots"
      },
      "post": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Create bot",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } }, "required": ["name"] } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/bots/{id}": {
      "delete": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Delete bot",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/api/admin/chats": {
      "get": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "List chats"
      },
      "post": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Create chat",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "chat_id": { "type": "string" }, "bot_id": { "type": "integer" } }, "required": ["name", "chat_id"] } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/chats/{id}": {
      "delete": {
        "tags": ["Admin"],
        "security": [{ "cookieAuth": [] }],
        "summary": "Delete chat",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
  }
}